+++ omnigibson/utils/ui_utils.py
@@ -8,7 +8,7 @@ import numpy as np
 import sys
 import datetime
 from pathlib import Path
-from PIL import Image
+from PIL import Image, ImageDraw
 from termcolor import colored
 import omnigibson as og
 from omnigibson.macros import gm
@@ -20,6 +20,7 @@ from scipy.integrate import quad
 import random
 import imageio
 from IPython import embed
+import colorsys
 
 
 def print_icon():
@@ -915,3 +916,55 @@ class KeyboardRobotController:
             print()
         print("*" * 30)
         print()
+
+def random_colours(N, enable_random=True, num_channels=3):
+    """
+    Generate random colors.
+    Generate visually distinct colours by linearly spacing the hue
+    channel in HSV space and then convert to RGB space.
+    """
+    start = 0
+    if enable_random:
+        random.seed(10)
+        start = random.random()
+    hues = [(start + i / N) % 1.0 for i in range(N)]
+    colours = [list(colorsys.hsv_to_rgb(h, 0.9, 1.0)) for i, h in enumerate(hues)]
+    if num_channels == 4:
+        for color in colours:
+            color.append(1.0)
+    if enable_random:
+        random.shuffle(colours)
+    return colours
+
+def colorize_bboxes(bboxes_2d_data, bboxes_2d_rgb, num_channels=3):
+    """Colorizes 2D bounding box data for visualization.
+
+
+    Args:
+        bboxes_2d_data (numpy.ndarray): 2D bounding box data from the sensor.
+        bboxes_2d_rgb (numpy.ndarray): RGB data from the sensor to embed bounding box.
+        num_channels (int): Specify number of channels i.e. 3 or 4.
+    """
+    semantic_id_list = []
+    bbox_2d_list = []
+    rgb_img = Image.fromarray(bboxes_2d_rgb)
+    rgb_img_draw = ImageDraw.Draw(rgb_img)
+    for bbox_2d in bboxes_2d_data:
+        semantic_id_list.append(bbox_2d['semanticId'])
+        bbox_2d_list.append(bbox_2d)
+    semantic_id_list_np = np.unique(np.array(semantic_id_list))
+    color_list = random_colours(len(semantic_id_list_np.tolist()), True, num_channels)
+    for bbox_2d in bbox_2d_list:
+        index = np.where(semantic_id_list_np == bbox_2d['semanticId'])[0][0]
+        bbox_color = color_list[index]
+        outline = (int(255 * bbox_color[0]), int(255 * bbox_color[1]), int(255 * bbox_color[2]))
+        if num_channels == 4:
+            outline = (
+                int(255 * bbox_color[0]),
+                int(255 * bbox_color[1]),
+                int(255 * bbox_color[2]),
+                int(255 * bbox_color[3]),
+            )
+        rgb_img_draw.rectangle([(bbox_2d['x_min'], bbox_2d['y_min']), (bbox_2d['x_max'], bbox_2d['y_max'])], outline=outline, width=2)
+    bboxes_2d_rgb = np.array(rgb_img)
+    return bboxes_2d_rgb
